eal/loongarch: add cpu cycle operations for LoongArch
This patch adds architecture specific cpu cycle operations for
LoongArch. The RDTIME.D instruction is used to read constant
frequency timer information including counter value. The CPUCFG
instruction is used to dynamically identify which features of
LoongArch are implemented in the running processor during the
execution of the software. We can use this instruction to calculate
the frequency used by the timer.

Signed-off-by: Min Zhou <zhoumin@loongson.cn>
Signed-off-by: 0-day Robot <robot@bytheb.org>