doc: update vhost sample app docs
Vhost sample app documentation describes parameters that are not in the
code and omits parameters that exist. Also switching the order of
sections on running vhost and VM, since the --client parameter in the
sample line requires a socket to be created by VM. Removing uio
references and updating with vfio-pci.

Signed-off-by: Herakliusz Lipiec <herakliusz.lipiec@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>