net/igc: move the initialization of data path into dev_init
The upper-layer application usually call the common interface "dev_init"
to initialize the data path, but in the igc driver, the initialization
of data path is in "igc_rx_init" and "eth_igc_tx_queue_setup", while in
other drivers it is in "dev_init". So when upper-layer application
calling these function pointers will occur segmentation faults.

This patch moves the initialization of data path into "eth_igc_dev_init"
to avoid segmentation faults, which is consistent with other drivers.

Fixes: a5aeb2b9e225 ("net/igc: support Rx and Tx")
Cc: alvinx.zhang@intel.com
Cc: stable@dpdk.org

Signed-off-by: Zhichao Zeng <zhichaox.zeng@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>