net/nfp: add initial flower firmware support
This commits adds the basic probing infrastructure to support the flower
firmware. This firmware is geared towards offloading OVS and can
generally be found in /lib/firmware/netronome/flower. It is also used by
the NFP kernel driver when OVS offload with TC is desired.

This commit also adds the basic infrastructure needed by the flower
firmware to operate. The firmware requires threads to service both the
PF vNIC and the ctrl vNIC. The PF is responsible for handling any
fallback traffic and the ctrl vNIC is used to communicate OVS flows
and flow statistics to and from the smartNIC. rte_services are used to
facilitate this logic.

This commit also adds the cpp service, used for some user tools.

Signed-off-by: Chaoyong He <chaoyong.he@corigine.com>
Signed-off-by: Heinrich Kuhn <heinrich.kuhn@corigine.com>
Reviewed-by: Niklas Söderlund <niklas.soderlund@corigine.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>